Add `EncodingContextProvider`
It is desirable to generalize the context provider of the encoding to a
map of string to string instead of requiring a disk read.

The filesystem is one of many providers of an encoding context, and it
is now available via a `EncodingContextFileProvider` struct.

This will facilitate the testing functions as well as make the encoding
process more flexible and customizable by the user.